TensorTrains.jl

Types

TensorTrains.AbstractTensorTrain — Type

AbstractTensorTrain

An abstract type representing a Tensor Train. Currently, there is only one concrete subtype TensorTrain.

TensorTrains.TensorTrain — Type

TensorTrain{F<:Number, N} <: AbstractTensorTrain

A type for representing a Tensor Train

F is the type of the matrix entries

N is the number of indices of each tensor (2 virtual ones + N-2 physical ones)

Functions

TensorTrains.normalize_eachmatrix! — Function

normalize_eachmatrix!(A::TensorTrain)

Divide each matrix by its maximum (absolute) element and return the sum of the logs of the individual normalizations. This is used to keep the entries from exploding during computations

TensorTrains.uniform_tt — Function

uniform_tt(bondsizes::AbstractVector{<:Integer}, q...)
uniform_tt(d::Integer, L::Integer, q...)

Construct a Tensor Train full of 1's, by specifying either:

bondsizes: the size of each bond

d a fixed size for all bonds, L the length

and

q a Tuple/Vector specifying the number of values taken by each variable on a single site

TensorTrains.rand_tt — Function

rand_tt(bondsizes::AbstractVector{<:Integer}, q...)
rand_tt(d::Integer, L::Integer, q...)

Construct a Tensor Train with entries random in [0,1], by specifying either:

bondsizes: the size of each bond

d a fixed size for all bonds, L the length

and

q a Tuple/Vector specifying the number of values taken by each variable on a single site

TensorTrains.bond_dims — Function

bond_dims(A::TensorTrain)

Return a vector with the dimensions of the virtual bonds

TensorTrains.evaluate — Function

evaluate(A::TensorTrain, X...)

Evaluate the Tensor Train A at input X

Example:

    L = 3
    q = (2, 3)
    A = rand_tt(4, L, q...)
    X = [[rand(1:qi) for qi in q] for l in 1:L]
    evaluate(A, X)

TensorTrains.marginals — Function

marginals(A::TensorTrain; l, r)

Compute the marginal distributions $p(x^l)$ at each site

Optional arguments

l = accumulate_L(A), r = accumulate_R(A) pre-computed partial nommalizations

TensorTrains.twovar_marginals — Function

marginals(A::TensorTrain; l, r, M, Δlmax)

Compute the marginal distributions for each pair of sites $p(x^l, x^m)$

Optional arguments

l = accumulate_L(A), r = accumulate_R(A), M = accumulate_M(A) pre-computed partial normalizations

Δlmax=length(A): compute marginals only at distance Δlmax: $|l-m|\le Δlmax$

TensorTrains.normalization — Function

normalization(A::TensorTrain; l, r)

Compute the normalization $Z=\sum_{x^1,\ldots,x^L} A^1(x^1)\cdots A^L(x^L)$

LinearAlgebra.normalize! — Method

normalize!(A::TensorTrain)

Normalize A to a probability distribution

Base.:+ — Function

+(A::TensorTrain, B::TensorTrain)

Compute the sum of two Tensor Trains. Matrix sizes are doubled

Base.:- — Function

-(A::TensorTrain, B::TensorTrain)

Compute the difference of two Tensor Trains. Matrix sizes are doubled

StatsBase.sample! — Function

sample!([rng], x, A::TensorTrain; r)

Draw an exact sample from A and store the result in x.

Optionally specify a random number generator rng as the first argument (defaults to Random.GLOBAL_RNG) and provide a pre-computed r = accumulate_R(A).

The output is x,p, the sampled sequence and its probability

StatsBase.sample — Function

sample([rng], A::TensorTrain; r)

Draw an exact sample from A.

Optionally specify a random number generator rng as the first argument (defaults to Random.GLOBAL_RNG) and provide a pre-computed r = accumulate_R(A).

The output is x,p, the sampled sequence and its probability