Types

TensorTrains.AbstractTensorTrain — Type

AbstractTensorTrain

An abstract type representing a Tensor Train. It currently supports 2 subtypes TensorTrain and PeriodicTensorTrain.

TensorTrains.TensorTrain — Type

TensorTrain{F<:Number, N} <: AbstractTensorTrain{F,N}

A type for representing a Tensor Train

F is the type of the matrix entries

N is the number of indices of each tensor (2 virtual ones + N-2 physical ones)

TensorTrains.PeriodicTensorTrain — Type

PeriodicTensorTrain{F<:Number, N} <: AbstractTensorTrain{F,N}

A type for representing a Tensor Train with periodic boundary conditions

F is the type of the matrix entries

N is the number of indices of each tensor (2 virtual ones + N-2 physical ones)

Functions

TensorTrains.normalize_eachmatrix! — Function

normalize_eachmatrix!(A::AbstractTensorTrain)

Divide each matrix by its maximum (absolute) element and return the sum of the logs of the individual normalizations. This is used to keep the entries from exploding during computations

TensorTrains.uniform_tt — Function

uniform_tt(bondsizes::AbstractVector{<:Integer}, q...)
uniform_tt(d::Integer, L::Integer, q...)

Construct a Tensor Train full of 1's, by specifying either:

bondsizes: the size of each bond

d a fixed size for all bonds, L the length

and

q a Tuple/Vector specifying the number of values taken by each variable on a single site

TensorTrains.rand_tt — Function

rand_tt(bondsizes::AbstractVector{<:Integer}, q...)
rand_tt(d::Integer, L::Integer, q...)

Construct a Tensor Train with entries random in [0,1], by specifying either:

bondsizes: the size of each bond

d a fixed size for all bonds, L the length

and

q a Tuple/Vector specifying the number of values taken by each variable on a single site

TensorTrains.uniform_periodic_tt — Function

uniform_periodic_tt(bondsizes::AbstractVector{<:Integer}, q...)
uniform_periodic_tt(d::Integer, L::Integer, q...)

Construct a Tensor Train with periodic boundary conditions full of 1's, by specifying either:

bondsizes: the size of each bond

d a fixed size for all bonds, L the length

and

q a Tuple/Vector specifying the number of values taken by each variable on a single site

TensorTrains.rand_periodic_tt — Function

rand_periodic_tt(bondsizes::AbstractVector{<:Integer}, q...)
rand_periodic_tt(d::Integer, L::Integer, q...)

Construct a Tensor Train with periodic boundary conditions with entries random in [0,1], by specifying either:

bondsizes: the size of each bond

d a fixed size for all bonds, L the length

and

q a Tuple/Vector specifying the number of values taken by each variable on a single site

TensorTrains.bond_dims — Function

bond_dims(A::AbstractTensorTrain)

Return a vector with the dimensions of the virtual bonds

TensorTrains.evaluate — Function

evaluate(A::AbstractTensorTrain, X...)

Evaluate the Tensor Train A at input X

Example:

    L = 3
    q = (2, 3)
    A = rand_tt(4, L, q...)
    X = [[rand(1:qi) for qi in q] for l in 1:L]
    evaluate(A, X)

TensorTrains.marginals — Function

marginals(A::AbstractTensorTrain; l, r)

Compute the marginal distributions $p(x^l)$ at each site

Optional arguments

l = accumulate_L(A), r = accumulate_R(A) pre-computed partial normalizations

TensorTrains.twovar_marginals — Function

twovar_marginals(A::AbstractTensorTrain; l, r, M, Δlmax)

Compute the marginal distributions for each pair of sites $p(x^l, x^m)$

Optional arguments

l = accumulate_L(A), r = accumulate_R(A), M = accumulate_M(A) pre-computed partial normalizations

maxdist = length(A): compute marginals only at distance maxdist: $|l-m|\le maxdist$

TensorTrains.normalization — Function

normalization(A::AbstractTensorTrain; l, r)

Compute the normalization $Z=\sum_{x^1,\ldots,x^L} A^1(x^1)\cdots A^L(x^L)$

LinearAlgebra.normalize! — Method

LinearAlgebra.normalize!(A::AbstractTensorTrain)

Normalize A to a probability distribution

Base.:+ — Function

Base.:(+)(A::AbstracTensorTrain, B::AbstracTensorTrain)

Compute the sum of two Tensor Trains. Matrix sizes are doubled

Base.:- — Function

Base.:(-)(A::AbstracTensorTrain, B::AbstracTensorTrain)

Compute the difference of two Tensor Trains. Matrix sizes are doubled

LinearAlgebra.dot — Function

LinearAlgebra.dot(A::AbstractTensorTrain, B::AbstractTensorTrain)

Compute the inner product between tensor trains A and B

\[A\cdot B = \sum_{x^1,x^2,\ldots,x^L}A^1(x^1)A^2(x^2)\cdots A^L(x^L)B^1(x^1)B^2(x^2)\cdots B^L(x^L)\]

LinearAlgebra.norm — Function

LinearAlgebra.norm(A::AbstractTensorTrain)

Compute the 2-norm (Frobenius norm) of tensor train A

\[\lVert A\rVert_2 = \sqrt{\sum_{x^1,x^2,\ldots,x^L}\left[A^1(x^1)A^2(x^2)\cdots A^L(x^L)\right]^2} = \sqrt{A\cdot A}\]

TensorTrains.norm2m — Function

norm2m(A::AbstractTensorTrain, B::AbstractTensorTrain)

Given two tensor trains A,B, compute norm(A - B)^2 as

\[\lVert A-B\rVert_2^2 = \lVert A \rVert_2^2 + \lVert B \rVert_2^2 - 2A\cdot B\]

StatsBase.sample! — Function

sample!([rng], x, A::AbstractTensorTrain; r)

Draw an exact sample from A and store the result in x.

Optionally specify a random number generator rng as the first argument (defaults to Random.GLOBAL_RNG) and provide a pre-computed r = accumulate_R(A).

The output is x,p, the sampled sequence and its probability

StatsBase.sample — Function

sample([rng], A::AbstractTensorTrain; r)

Draw an exact sample from A.

Optionally specify a random number generator rng as the first argument (defaults to Random.GLOBAL_RNG) and provide a pre-computed r = accumulate_R(A).

The output is x,p, the sampled sequence and its probability

TensorTrains.orthogonalize_right! — Function

orthogonalize_right!(A::AbstractTensorTrain; svd_trunc::SVDTrunc)

Bring A to right-orthogonal form by means of SVD decompositions.

Optionally perform truncations by passing a SVDTrunc.

TensorTrains.orthogonalize_left! — Function

orthogonalize_left!(A::AbstractTensorTrain; svd_trunc::SVDTrunc)

Bring A to left-orthogonal form by means of SVD decompositions.

Optionally perform truncations by passing a SVDTrunc.

TensorTrains.compress! — Function

compress!(A::AbstractTensorTrain; svd_trunc::SVDTrunc)

Compress A by means of SVD decompositions + truncations

SVD Truncators

TensorTrains.SVDTrunc — Type

abstract type SVDTrunc

SVD truncator. Can be threshold-based or bond size-based

TensorTrains.TruncThresh — Type

TruncThresh{T} <: SVDTrunc

A type used to perform SVD-based truncations based on a threshold ε. Given a vector $\lambda$ of $m$ singular values, those below $\varepsilon\sqrt{\sum_{k=1}^m \lambda_k^2}$ are truncated to zero.

FIELDS

ε: threshold.

svd_trunc = TruncThresh(1e-5)
M = rand(5,6)
M_trunc = svd_trunc(M)

TensorTrains.TruncBond — Type

TruncBond{T} <: SVDTrunc

A type used to perform SVD-based truncations based on bond size m'. Given a vector $\lambda$ of $m$ singular values, only the $m'$ largest are kept, the others are truncated to zero.

FIELDS

mprime: number of singular values to retain

svd_trunc = TruncBond(3)
M = rand(5,6)
M_trunc = svd_trunc(M)

TensorTrains.TruncBondMax — Type

TruncBondMax{T} <: SVDTrunc

Similar to TruncBond, but also stores the maximum error $\sqrt{\frac{\sum_{k=m'+1}^m\lambda_k^2}{\sum_{k=1}^m\lambda_k^2}}$ made since the creation of the object

FIELDS

mprime: number of singular values to retain

maxerr: a 1-element vector storing the maximum error

TensorTrains.TruncBondThresh — Type

TruncBondThresh{T} <: SVDTrunc

A mixture of TruncBond and TruncThresh, truncates to the most stringent criterion.